backend/requirements.txt
celery==4.2.1
Django==2.1.1
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
django-redis-sessions==0.6.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
flake8==3.5.0
raven==6.9.0
urlparse3==1.1
pytest-django==3.4.5
pytest-factoryboy==2.0.2